# docs/swagger.yaml
definitions:
...
  CreateUserVO:
    properties:
      department:
        description: 部门
        type: string
      loginName:
        description: 工号
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码，rsa加密
        type: string
    required:
    - department
    - loginName
    - name
    - password
    type: object
...
  DeleteVO:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    type: object
...
  EncryptPassWordVO:
    properties:
      plaintext:
        description: 明文密码
        type: string
    required:
    - plaintext
    type: object
...
  LoginVO:
    properties:
      loginName:
        description: 工号
        type: string
      passWord:
        description: 密码，rsa加密
        type: string
    required:
    - loginName
    - passWord
    type: object
...
  ModifyPasswordVO:
    properties:
      newPwd:
        description: 新密码，rsa加密
        type: string
      oldPwd:
        description: 原始密码，rsa加密
        type: string
    required:
    - newPwd
    - oldPwd
    type: object
...
  UpdateUserAuthorityVO:
    properties:
      codes:
        description: 权限编码
        items:
          type: string
        minItems: 1
        type: array
      loginName:
        description: 工号
        type: string
    required:
    - loginName
    type: object
  UpdateUserVO:
    properties:
      department:
        description: 部门
        type: string
      id:
        description: ID
        type: integer
      loginName:
        description: 工号
        type: string
      name:
        description: 姓名
        type: string
    required:
    - department
    - id
    - loginName
    - name
    type: object
...
info:
  contact: {}
  description: 模具管理API
  title: 模具管理API
  version: "1.0"
paths:
...
  /user:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/DeleteVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 姓名或者工号
        in: query
        name: name
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateUserVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 添加用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/UpdateUserVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 更新用户
      tags:
      - 用户管理
  /user/authorities:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 获取所有可配置的权限
      tags:
      - 权限
  /user/authority:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 获取当前登录用户权限编码
      tags:
      - 权限
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/UpdateUserAuthorityVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 更新指定用户权限
      tags:
      - 权限
  /user/authority/one:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户登录名
        in: query
        name: loginName
        required: true
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 获取指定用户权限编码,管理员only
      tags:
      - 权限
  /user/encrypt:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/EncryptPassWordVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 密码加密
      tags:
      - 登录
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/LoginVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - 登录
  /user/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 登出当前用户
      tags:
      - 登录
  /user/one:
    get:
      consumes:
      - application/json
      parameters:
      - description: loginName
        in: query
        name: loginName
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 查询单个用户信息,管理员only
      tags:
      - 用户管理
  /user/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 姓名或者工号
        in: query
        name: name
        type: string
      - description: 当前页
        in: query
        name: currentPage
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 用户分页
      tags:
      - 用户管理
  /user/pwd:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/ModifyPasswordVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 修改当前用户密码
      tags:
      - 用户管理
  /user/reset:
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/LoginVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 重置密码,only管理员
      tags:
      - 用户管理
  /user/rsa:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 获取rsa公钥
      tags:
      - 登录
  /user/self:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 查询当前登录用户信息
      tags:
      - 用户管理
swagger: "2.0"